refactor code split logic into fragments
