+++ app/src/main/res/menu/main_menu.xml
@@ -6,4 +6,11 @@
         android:title="@string/about"
         android:icon="@drawable/ic_baseline_info_24"
         app:showAsAction="always"/>
+
+    <item
+        android:id="@+id/setting"
+        android:title="@string/about"
+        android:icon="@drawable/ic_settings"
+        app:showAsAction="always"/>
+
 </menu>
